+++ src/main/scala/com/kotori316/fluidtank/fluids/GenericAmount.scala
@@ -3,7 +3,7 @@ package com.kotori316.fluidtank.fluids
 import java.util.Objects
 
 import cats.implicits.{catsSyntaxEq, catsSyntaxGroup, catsSyntaxSemigroup}
-import cats.{Hash, Show}
+import cats.{Hash, Monoid, Show}
 import com.kotori316.fluidtank.eqCompoundNbt
 import net.minecraft.nbt.CompoundTag
 
@@ -66,7 +66,27 @@ class GenericAmount[Content]
   }
 }
 
-object GenericAmount {
+object GenericAmount extends FluidAmountImplicits with GeneralImplicits {
+
+}
+
+private trait FluidAmountImplicits {
+  implicit val showFA: Show[FluidAmount] = Show.fromToString
+  implicit val hashFA: Hash[FluidAmount] = new Hash[FluidAmount] {
+    override def hash(x: FluidAmount): Int = x.hashCode()
+
+    override def eqv(x: FluidAmount, y: FluidAmount): Boolean = x.fluidEqual(y) && (x.amount === y.amount)
+  }
+
+  implicit val monoidFA: Monoid[FluidAmount] = new Monoid[FluidAmount] {
+    override def empty: FluidAmount = FluidAmount.EMPTY
+
+    override def combine(x: FluidAmount, y: FluidAmount): FluidAmount = x + y
+  }
+
+}
+
+private trait GeneralImplicits {
   implicit def showGA[A]: Show[GenericAmount[A]] = Show.fromToString
 
   implicit def hashGA[A]: Hash[GenericAmount[A]] = new Hash[GenericAmount[A]] {
